requirements.txt
django==3.2.2
djangorestframework==3.12.4
psycopg2==2.8.6
flake8==3.9.2
 Pillow==8.2.0